zombies spawn in the correct position over the mouse

+++ Andrew_Test/Assets/UI/Picker/ZombiePickerMaster.js
@@ -2,8 +2,8 @@
 
 var zombies_arr:GameObject[];
 static var zombieIndex:int=-1;
-static var mouse_x:int=0;
-static var mouse_y:int=0;
+static var mouse_x:float=0;
+static var mouse_y:float=0;
 var  uiCamera:Camera;
